Explicit retries in ImageStoreServiceWrapper Retry ESExporter.index_ds on ConnectionTimeout
